refactor(VendingMachineController): dependency injection structure
