+++ src/main/java/vendingmachine/Application.java
@@ -1,11 +1,11 @@
 package vendingmachine;
 
-import vendingmachine.domain.VendingMachine;
+import vendingmachine.controller.VendingMachineController;
 
 public class Application {
 
     public static void main(String[] args) {
-        VendingMachine machine = new VendingMachine();
+        VendingMachineController machine = new VendingMachineController();
         machine.run();
     }
 }
